set barcode bugg & fix crop scan size
I noticed that setting barcode types happend 1 time and was triggered before the properties was passed from react. So it was always scanning all possible formats.

Also I've changed the crop size to be based on % size. This because different resolutions on cameras didn't match up with size from react.
There's an commented code section when capture image that can be used to tweak the cropsize, so we know what we actually are looking at.