Add iOS support for start/stop preview in RNCamera
This is based on the work in #859. It adds only the iOS version of these
because the Android API is now based on the preview of
`android.cameraview.CameraView`
([fork](https://github.com/react-native-community/cameraview)) rather
than the prior custom implementation. Hopefully someone else can work
through adding this in the future.

It also adds a flag to the `takePictureAsync` to automatically stop the
preview immediately once the still capture completes (and before any
additional async activity occurs).

Test environment:

iOS 11.2.5
React Native 0.54.2
Node 9.5.0