Fix the git tag format in the podspec
The format of the git tag which the podspec file specified did not match the ones which were being published.

Previously, it was using the raw version number from the `package.json`, however, the git tags had a `v` prefix.

This lead to errors saying the "branch could not be found" when attempting to use the library through Cocoapods.

This fix still reads the version number from `package.json`, however, it now prefixes it with a `v` to get the real git tag values. For example, `v1.6.4`.