feat(rn-camera): add ActivityIndicator style prop
I found no clue of how to style the RNCamera's activity indicator, so I saw that no props were passed to the component.

I removed the PropType element pendingAuthorizationView and created a
function called renderPendingAuthorizationView.
Also added activityIndicatorStyle to RNCameraProps in index.d.ts and
used the prop in the example app.

Now we can use activityIndicatorStyle, an object with color and size
keys, as a RNCamera prop.